Fix that single godimages apply to all hand cards
It is possible to request a godimage when more than one copy of the card
exists in a player's hand.
The code updated the godimage for *all* copies of the card for which a
godimage was requested in the hand of the player that made the request.
This commit fixes that behavior so that the single godimage is only
applied to the card at God rarity.

This would allow the creation of Mythical and below cards with the single
godimage.
The lower rarity godimaged cards are non-transferable due to the trade
code doing an explicit rarity check on transfer and bounties cause the
creation of a new card anyway.